refactoring forgoten code
