+++ go.mod
@@ -4,5 +4,6 @@ go 1.15
 
 replace (
 	lucasmontano.com/yt-links/models => ./models
+	lucasmontano.com/yt-links/services => ./services
 	lucasmontano.com/yt-links/tests => ./tests
 )
