+++ services/description_to_links_mapper.go
@@ -1,4 +1,4 @@
-package main
+package services
 
 import "regexp"
